requirements.txt
# AI Transformation Discovery Agent — Python Dependencies
# Pin exact versions to prevent drift between environments

# Core framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0

# HTTP client for scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.0

# Model providers
google-cloud-aiplatform==1.64.0    # Vertex AI (primary)
anthropic==0.34.0                   # Evals judge only
google-auth>=2.29.0                 # AnthropicVertex GCP auth

# Vector store (local for Sprint 1)
chromadb==0.5.0
numpy<2.0                               # ChromaDB 0.5 incompatible with numpy 2.x

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.2
tenacity==8.5.0                     # Retry logic
